requirements.txt
Flask
pymongo
Flask-Login
Flask-Bcrypt
Flask-Cors
python-dotenv
pytz
requests
python-docx
pandas
Werkzeug
jwt
smtplib